+++ ChangeLog.md
@@ -1,5 +1,14 @@
 # Revision history for mattermost-api-qc
 
+## 30802.1.0  -- 2017-05-13
+
+* Switched from Arbitrary instances to Gen x functions.
+* Removed dependency on quickcheck-instances
+
+## 30802.0.0  -- 2017-05-11
+
+* Updated for mattermost-api version 30802.0.0
+	
 ## 30701.1.0  -- 2017-05-09
 
 * First version.